\name{ct.possible.models}
\alias{ct.possible.models}
\title{Count all possible model combinations}
\description{
Count all combinations of predictor variables in a multivariate regression model.
}
\usage{
ct.possible.models(q)
}
\arguments{
\item{q}{number of predictor variables}
}
\value{
a count of the number of possible models
}
\examples{

ct.possible.models(9)

}
